Move more reporters under crier
