+++ prow/BUILD.bazel
@@ -144,7 +144,6 @@ filegroup(
         "//prow/pod-utils/options:all-srcs",
         "//prow/pod-utils/wrapper:all-srcs",
         "//prow/prstatus:all-srcs",
-        "//prow/pubsub/reporter:all-srcs",
         "//prow/pubsub/subscriber:all-srcs",
         "//prow/repoowners:all-srcs",
         "//prow/sidecar:all-srcs",
