+++ prow/cmd/crier/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//prow/crier/reporters/gcs:go_default_library",
         "//prow/crier/reporters/gerrit:go_default_library",
         "//prow/crier/reporters/github:go_default_library",
+        "//prow/crier/reporters/pubsub:go_default_library",
         "//prow/crier/reporters/slack:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/gerrit/client:go_default_library",
@@ -22,7 +23,6 @@ go_library(
         "//prow/kube:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/pjutil:go_default_library",
-        "//prow/pubsub/reporter:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
         "@org_golang_google_api//option:go_default_library",
