+++ prow/cmd/sub/BUILD.bazel
@@ -11,11 +11,11 @@ go_library(
         "//prow/client/clientset/versioned/typed/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
         "//prow/config/secret:go_default_library",
+        "//prow/crier/reporters/pubsub:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/interrupts:go_default_library",
         "//prow/logrusutil:go_default_library",
         "//prow/metrics:go_default_library",
-        "//prow/pubsub/reporter:go_default_library",
         "//prow/pubsub/subscriber:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
