+++ prow/cmd/sub/main.go
@@ -31,10 +31,10 @@ import (
 	prowv1 "k8s.io/test-infra/prow/client/clientset/versioned/typed/prowjobs/v1"
 	"k8s.io/test-infra/prow/config"
 	"k8s.io/test-infra/prow/config/secret"
+	"k8s.io/test-infra/prow/crier/reporters/pubsub"
 	"k8s.io/test-infra/prow/flagutil"
 	"k8s.io/test-infra/prow/logrusutil"
 	"k8s.io/test-infra/prow/metrics"
-	"k8s.io/test-infra/prow/pubsub/reporter"
 	"k8s.io/test-infra/prow/pubsub/subscriber"
 )
 
@@ -125,7 +125,7 @@ func main() {
 		ConfigAgent:   configAgent,
 		Metrics:       promMetrics,
 		ProwJobClient: kubeClient,
-		Reporter:      reporter.NewReporter(configAgent.Config),
+		Reporter:      pubsub.NewReporter(configAgent.Config),
 	}
 
 	// Return 200 on / for health checks.
