+++ prow/crier/OWNERS
@@ -2,5 +2,6 @@
 
 approvers:
 - krzyzacy
+- katharine
 labels:
 - area/prow/crier
