+++ prow/pubsub/subscriber/BUILD.bazel
@@ -29,7 +29,7 @@ go_test(
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/client/clientset/versioned/fake:go_default_library",
         "//prow/config:go_default_library",
-        "//prow/pubsub/reporter:go_default_library",
+        "//prow/crier/reporters/pubsub:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go//pubsub:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
