+++ prow/pubsub/subscriber/subscriber_test.go
@@ -37,7 +37,7 @@ import (
 	prowapi "k8s.io/test-infra/prow/apis/prowjobs/v1"
 	"k8s.io/test-infra/prow/client/clientset/versioned/fake"
 	"k8s.io/test-infra/prow/config"
-	"k8s.io/test-infra/prow/pubsub/reporter"
+	reporter "k8s.io/test-infra/prow/crier/reporters/pubsub"
 
 	v1 "k8s.io/api/core/v1"
 )
